CoreTransaction & ICoreTransaction
