+++ src/model2/transaction.ts
@@ -1,23 +1,11 @@
-import stellar from "stellar-base";
 import { Memo } from "stellar-sdk";
-import { publicKeyFromBuffer } from "../util/xdr";
 
 export type TimeBounds = [number, number];
 
-// NOTE!!! ITransaction and Transaction should implement interface for GraphQL only
-// CoreTransaction must have all that XDR stuff, reiterate here.
 export interface ITransaction {
   id: string;
   ledgerSeq: number;
   index: number;
-  body?: string;
-  bodyXDR?: any;
-  result?: string;
-  resultXDR?: any;
-  meta?: string;
-  metaXDR?: any;
-  feeMeta?: string;
-  feeMetaXDR?: any;
   memo?: Memo;
   feeAmount: string;
   sourceAccount: string;
@@ -27,74 +15,10 @@ export interface ITransaction {
   resultCode: number;
 }
 
-export interface ITransactionTableRow {
-  txid: string;
-  ledgerseq: number;
-  txindex: number;
-  txbody: string;
-  txresult: string;
-  txmeta: string;
-  txfeemeta: string;
-}
-
 export class Transaction implements ITransaction {
-  public static fromDb(row: ITransactionTableRow): Transaction {
-    const bodyXDR = stellar.xdr.TransactionEnvelope.fromXDR(Buffer.from(row.txbody, "base64"));
-    const resultXDR = stellar.xdr.TransactionEnvelope.fromXDR(Buffer.from(row.txresult, "base64"));
-    const metaXDR = stellar.xdr.TransactionMeta.fromXDR(Buffer.from(row.txmeta, "base64"));
-    const feeMetaXDR = stellar.xdr.OperationMeta.fromXDR(Buffer.from(row.txfeemeta, "base64"));
-
-    const body = bodyXDR.tx();
-    const result = resultXDR.result();
-
-    const memo = Memo.fromXDRObject(body.memo());
-    const timeBoundsXDR = body.timeBounds();
-
-    const timeBounds: TimeBounds | undefined = timeBoundsXDR
-      ? [timeBoundsXDR.minTime().toInt(), timeBoundsXDR.maxTime().toInt()]
-      : undefined;
-
-    const resultCode = result.result().switch().value;
-    const success = resultCode === stellar.xdr.TransactionResultCode.txSuccess().value;
-    const feeAmount = body.fee().toString();
-    const feeCharged = result.feeCharged.toString();
-    const sourceAccount = publicKeyFromBuffer(body.sourceAccount().value());
-
-    const data: ITransaction = {
-      id: row.txid,
-      ledgerSeq: row.ledgerseq,
-      index: row.txindex,
-      body: row.txbody,
-      bodyXDR,
-      result: row.txresult,
-      resultXDR,
-      meta: row.txmeta,
-      metaXDR,
-      feeMeta: row.txfeemeta,
-      feeMetaXDR,
-      memo: memo.value ? memo : undefined,
-      timeBounds,
-      feeAmount,
-      feeCharged,
-      resultCode,
-      success,
-      sourceAccount
-    };
-
-    return new Transaction(data);
-  }
-
   public id: string;
   public ledgerSeq: number;
   public index: number;
-  public body?: string;
-  public bodyXDR?: any;
-  public result?: string;
-  public resultXDR?: any;
-  public meta?: string;
-  public metaXDR?: any;
-  public feeMeta?: string;
-  public feeMetaXDR?: any;
   public memo?: Memo;
   public feeAmount: string;
   public sourceAccount: string;
@@ -107,14 +31,6 @@ export class Transaction implements ITransaction {
     this.id = data.id;
     this.ledgerSeq = data.ledgerSeq;
     this.index = data.index;
-    this.body = data.body;
-    this.bodyXDR = data.bodyXDR;
-    this.result = data.result;
-    this.resultXDR = data.resultXDR;
-    this.meta = data.meta;
-    this.metaXDR = data.metaXDR;
-    this.feeMeta = data.feeMeta;
-    this.feeMetaXDR = data.feeMetaXDR;
     this.memo = data.memo;
     this.feeAmount = data.feeAmount;
     this.sourceAccount = data.sourceAccount;
